requirements.txt
art==6.1
asgiref==3.7.2
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
chromedriver-py==125.0.6422.60
click==8.1.7
crispy-bootstrap5==2024.10
cs50==9.3.4
dj-database-url==2.3.0
Django==5.0.2
django-cors-headers==4.6.0
django-crispy-forms==2.3
Flask==3.0.0
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
Markdown==3.7
markdown2==2.4.13
MarkupSafe==2.1.3
outcome==1.3.0.post0
packaging==23.2
pillow==11.0.0
psycopg2-binary==2.9.10
PySocks==1.7.1
python-dotenv==1.0.1
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
termcolor==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.1
wheel==0.42.0
whitenoise==6.8.2
wsproto==1.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
